// packages/academic/src/application/use-cases/supplement-journals-with-semantic-scholar/supplement-journals-with-semantic-scholar.dto.ts
export interface SupplementJournalsWithSemanticScholarOutput {
  journalsClaimed: number;
  journalsCompleted: number;
  journalsShortfall: number;
  bulkPagesFetched: number;
  recommendationCalls: number;
  newAccepted: number;
  relatedAccepted: number;
  articlesInserted: number;
  articlesUpdated: number;
}